+++ packages/example/pages/_app.tsx
@@ -45,6 +45,7 @@ import {
   tahoWallet,
   talismanWallet,
   tokenPocketWallet,
+  tokenaryWallet,
   trustWallet,
   uniswapWallet,
   xdefiWallet,