+++ packages/rainbow-button/src/components/RainbowButton.tsx
@@ -1,0 +1,24 @@
+import { RainbowKitProvider, WalletButton } from '@rainbow-me/rainbowkit';
+import type { RainbowKitProviderProps } from '@rainbow-me/rainbowkit/dist/components/RainbowKitProvider/RainbowKitProvider';
+import React from 'react';
+
+export function RainbowButtonProvider({
+  children,
+  ...options
+}: Omit<
+  RainbowKitProviderProps,
+  'chains' | 'avatar' | 'initialChain' | 'modalSize' | 'showRecentTransactions'
+>) {
+  return (
+    // chains array is intentionally left empty
+    <RainbowKitProvider chains={[]} {...options}>
+      {children}
+    </RainbowKitProvider>
+  );
+}
+
+export const RainbowButton = () => {
+  return <WalletButton wallet="rainbow" />;
+};
+
+RainbowButton.Custom = WalletButton.Custom;