+++ packages/rainbow-button/src/index.ts
@@ -1,0 +1,3 @@
+export { RainbowButtonProvider } from './components/RainbowButton';
+export { RainbowConnector } from './connectors/rainbow';
+export { RainbowButton } from './components/RainbowButton';