+++ packages/rainbowkit/src/components/ConnectOptions/MobileOptions.tsx
@@ -1,30 +1,68 @@
-import React, { useCallback, useContext, useState } from 'react';
-import { touchableStyles } from '../../css/touchableStyles';
-import { isIOS } from '../../utils/isMobile';
+import React, {
+  useCallback,
+  useContext,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
+import { touchableStyles } from "../../css/touchableStyles";
+import { isIOS } from "../../utils/isMobile";
 import {
   WalletConnector,
   useWalletConnectors,
-} from '../../wallets/useWalletConnectors';
-import { AsyncImage } from '../AsyncImage/AsyncImage';
-import { Box } from '../Box/Box';
-import { ActionButton } from '../Button/ActionButton';
-import { CloseButton } from '../CloseButton/CloseButton';
-import { DisclaimerLink } from '../Disclaimer/DisclaimerLink';
-import { DisclaimerText } from '../Disclaimer/DisclaimerText';
-import { BackIcon } from '../Icons/Back';
-import { AppContext } from '../RainbowKitProvider/AppContext';
-import { I18nContext } from '../RainbowKitProvider/I18nContext';
-import { useCoolMode } from '../RainbowKitProvider/useCoolMode';
-import { setWalletConnectDeepLink } from '../RainbowKitProvider/walletConnectDeepLink';
-import { Text } from '../Text/Text';
-import * as styles from './MobileOptions.css';
+} from "../../wallets/useWalletConnectors";
+import { AsyncImage } from "../AsyncImage/AsyncImage";
+import { Box } from "../Box/Box";
+import { ActionButton } from "../Button/ActionButton";
+import { CloseButton } from "../CloseButton/CloseButton";
+import { DisclaimerLink } from "../Disclaimer/DisclaimerLink";
+import { DisclaimerText } from "../Disclaimer/DisclaimerText";
+import { BackIcon } from "../Icons/Back";
+import { AppContext } from "../RainbowKitProvider/AppContext";
+import { I18nContext } from "../RainbowKitProvider/I18nContext";
+import { useCoolMode } from "../RainbowKitProvider/useCoolMode";
+import { setWalletConnectDeepLink } from "../RainbowKitProvider/walletConnectDeepLink";
+import { Text } from "../Text/Text";
+import * as styles from "./MobileOptions.css";
 
-function WalletButton({
+const LoadingSpinner = ({ wallet }: { wallet: WalletConnector }) => {
+  const width = 80;
+  const height = 80;
+  const radiusFactor = 20;
+
+  const perimeter = 2 * (width + height - 4 * radiusFactor);
+
+  return (
+    <svg className={styles.spinner} viewBox="0 0 86 86" width="86" height="86">
+      <rect
+        x="3"
+        y="3"
+        width={width}
+        height={height}
+        rx={radiusFactor}
+        ry={radiusFactor}
+        strokeDasharray={`${perimeter / 3} ${(2 * perimeter) / 3}`}
+        strokeDashoffset={perimeter} // Adjust this value as per your design needs
+        className={styles.rotatingBorder}
+        style={{
+          // Prop style passing works only in `@vanilla-extract/recipes`.
+          // Instead downloading packages we can do this
+          // manually without passing props
+          stroke: wallet?.iconAccent || "#0D3887",
+        }}
+      />
+    </svg>
+  );
+};
+
+export function WalletButton({
   onClose,
   wallet,
+  connecting,
 }: {
   wallet: WalletConnector;
   onClose: () => void;
+  connecting?: boolean;
 }) {
   const {
     connect,
@@ -39,58 +77,71 @@ function WalletButton({
   } = wallet;
 
   const coolModeRef = useCoolMode(iconUrl);
+  const initialized = useRef(false);
 
   const i18n = useContext(I18nContext);
+
+  const onConnect = useCallback(async () => {
+    const onMobileUri = async () => {
+      const mobileUri = await getMobileUri?.();
+
+      if (!mobileUri) return;
+
+      if (mobileUri) {
+        setWalletConnectDeepLink({ mobileUri, name });
+      }
+
+      if (mobileUri.startsWith("http")) {
+        // Workaround for https://github.com/rainbow-me/rainbowkit/issues/524.
+        // Using 'window.open' causes issues on iOS in non-Safari browsers and
+        // WebViews where a blank tab is left behind after connecting.
+        // This is especially bad in some WebView scenarios (e.g. following a
+        // link from Twitter) where the user doesn't have any mechanism for
+        // closing the blank tab.
+        // For whatever reason, links with a target of "_blank" don't suffer
+        // from this problem, and programmatically clicking a detached link
+        // element with the same attributes also avoids the issue.
+        const link = document.createElement("a");
+        link.href = mobileUri;
+        link.target = "_blank";
+        link.rel = "noreferrer noopener";
+        link.click();
+      } else {
+        window.location.href = mobileUri;
+      }
+    };
+
+    if (id !== "walletConnect") onMobileUri();
+
+    if (showWalletConnectModal) {
+      showWalletConnectModal();
+      onClose?.();
+      return;
+    }
+
+    connect?.();
+  }, [connect, getMobileUri, showWalletConnectModal, onClose, name, id]);
+
+  // biome-ignore lint/nursery/useExhaustiveDependencies: <explanation>
+  useEffect(() => {
+    // When using `reactStrictMode: true` in development mode the useEffect hook
+    // will fire twice. We avoid this by using `useRef` logic here. Works for now.
+    if (connecting && !initialized.current) {
+      onConnect();
+      initialized.current = true;
+    }
+  }, [connecting]);
 
   return (
     <Box
       as="button"
-      color={ready ? 'modalText' : 'modalTextSecondary'}
+      color={ready ? "modalText" : "modalTextSecondary"}
       disabled={!ready}
       fontFamily="body"
       key={id}
-      onClick={useCallback(async () => {
-        const onMobileUri = async () => {
-          const mobileUri = await getMobileUri?.();
-
-          if (!mobileUri) return;
-
-          if (mobileUri) {
-            setWalletConnectDeepLink({ mobileUri, name });
-          }
-
-          if (mobileUri.startsWith('http')) {
-            // Workaround for https://github.com/rainbow-me/rainbowkit/issues/524.
-            // Using 'window.open' causes issues on iOS in non-Safari browsers and
-            // WebViews where a blank tab is left behind after connecting.
-            // This is especially bad in some WebView scenarios (e.g. following a
-            // link from Twitter) where the user doesn't have any mechanism for
-            // closing the blank tab.
-            // For whatever reason, links with a target of "_blank" don't suffer
-            // from this problem, and programmatically clicking a detached link
-            // element with the same attributes also avoids the issue.
-            const link = document.createElement('a');
-            link.href = mobileUri;
-            link.target = '_blank';
-            link.rel = 'noreferrer noopener';
-            link.click();
-          } else {
-            window.location.href = mobileUri;
-          }
-        };
-
-        if (id !== 'walletConnect') onMobileUri();
-
-        if (showWalletConnectModal) {
-          showWalletConnectModal();
-          onClose?.();
-          return;
-        }
-
-        connect?.();
-      }, [connect, getMobileUri, showWalletConnectModal, onClose, name, id])}
+      onClick={onConnect}
       ref={coolModeRef}
-      style={{ overflow: 'visible', textAlign: 'center' }}
+      style={{ overflow: "visible", textAlign: "center" }}
       testId={`wallet-option-${id}`}
       type="button"
       width="full"
@@ -101,7 +152,15 @@ function WalletButton({
         flexDirection="column"
         justifyContent="center"
       >
-        <Box paddingBottom="8" paddingTop="10">
+        <Box
+          display="flex"
+          alignItems="center"
+          justifyContent="center"
+          paddingBottom="8"
+          paddingTop="10"
+          position="relative"
+        >
+          {connecting ? <LoadingSpinner wallet={wallet} /> : null}
           <AsyncImage
             background={iconBackground}
             borderRadius="13"
@@ -111,40 +170,42 @@ function WalletButton({
             width="60"
           />
         </Box>
-        <Box display="flex" flexDirection="column" textAlign="center">
-          <Text
-            as="h2"
-            color={wallet.ready ? 'modalText' : 'modalTextSecondary'}
-            size="13"
-            weight="medium"
-          >
-            {/* Fix button text clipping in Safari: https://stackoverflow.com/questions/41100273/overflowing-button-text-is-being-clipped-in-safari */}
-            <Box as="span" position="relative">
-              {shortName ?? name}
-              {!wallet.ready && ' (unsupported)'}
-            </Box>
-          </Text>
-
-          {wallet.recent && (
-            <Text color="accentColor" size="12" weight="medium">
-              {i18n.t('connect.recent')}
+        {!connecting ? (
+          <Box display="flex" flexDirection="column" textAlign="center">
+            <Text
+              as="h2"
+              color={wallet.ready ? "modalText" : "modalTextSecondary"}
+              size="13"
+              weight="medium"
+            >
+              {/* Fix button text clipping in Safari: https://stackoverflow.com/questions/41100273/overflowing-button-text-is-being-clipped-in-safari */}
+              <Box as="span" position="relative">
+                {shortName ?? name}
+                {!wallet.ready && " (unsupported)"}
+              </Box>
             </Text>
-          )}
-        </Box>
+
+            {wallet.recent && (
+              <Text color="accentColor" size="12" weight="medium">
+                {i18n.t("connect.recent")}
+              </Text>
+            )}
+          </Box>
+        ) : null}
       </Box>
     </Box>
   );
 }
 
 enum MobileWalletStep {
-  Connect = 'CONNECT',
-  Get = 'GET',
+  Connect = "CONNECT",
+  Get = "GET",
 }
 
 export function MobileOptions({ onClose }: { onClose: () => void }) {
-  const titleId = 'rk_connect_title';
+  const titleId = "rk_connect_title";
   const wallets = useWalletConnectors().filter(
-    (wallet) => wallet.isRainbowKitConnector,
+    (wallet) => wallet.isRainbowKitConnector
   );
   const { disclaimer: Disclaimer, learnMoreUrl } = useContext(AppContext);
 
@@ -154,7 +215,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
   let headerBackButtonLink: MobileWalletStep | null = null;
 
   const [walletStep, setWalletStep] = useState<MobileWalletStep>(
-    MobileWalletStep.Connect,
+    MobileWalletStep.Connect
   );
 
   const i18n = useContext(I18nContext);
@@ -163,7 +224,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
 
   switch (walletStep) {
     case MobileWalletStep.Connect: {
-      headerLabel = i18n.t('connect.title');
+      headerLabel = i18n.t("connect.title");
       headerBackgroundContrast = true;
       walletContent = (
         <Box>
@@ -174,7 +235,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
             paddingBottom="20"
             paddingTop="6"
           >
-            <Box display="flex" style={{ margin: '0 auto' }}>
+            <Box display="flex" style={{ margin: "0 auto" }}>
               {wallets
                 .filter((wallet) => wallet.ready)
                 .map((wallet) => {
@@ -202,7 +263,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
             flexDirection="column"
             gap="32"
             paddingX="32"
-            style={{ textAlign: 'center' }}
+            style={{ textAlign: "center" }}
           >
             <Box
               display="flex"
@@ -211,10 +272,10 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
               textAlign="center"
             >
               <Text color="modalText" size="16" weight="bold">
-                {i18n.t('intro.title')}
+                {i18n.t("intro.title")}
               </Text>
               <Text color="modalTextSecondary" size="16">
-                {i18n.t('intro.description')}
+                {i18n.t("intro.description")}
               </Text>
             </Box>
           </Box>
@@ -222,14 +283,14 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
           <Box paddingTop="32" paddingX="20">
             <Box display="flex" gap="14" justifyContent="center">
               <ActionButton
-                label={i18n.t('intro.get.label')}
+                label={i18n.t("intro.get.label")}
                 onClick={() => setWalletStep(MobileWalletStep.Get)}
                 size="large"
                 type="secondary"
               />
               <ActionButton
                 href={learnMoreUrl}
-                label={i18n.t('intro.learn_more.label')}
+                label={i18n.t("intro.learn_more.label")}
                 size="large"
                 type="secondary"
               />
@@ -245,7 +306,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
       break;
     }
     case MobileWalletStep.Get: {
-      headerLabel = i18n.t('get.title');
+      headerLabel = i18n.t("get.title");
       headerBackButtonLink = MobileWalletStep.Connect;
 
       const mobileWallets = wallets
@@ -253,7 +314,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
           (wallet) =>
             wallet.downloadUrls?.ios ||
             wallet.downloadUrls?.android ||
-            wallet.downloadUrls?.mobile,
+            wallet.downloadUrls?.mobile
         )
         ?.splice(0, 3);
 
@@ -310,7 +371,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
                           (ios ? downloadUrls?.ios : downloadUrls?.android) ||
                           downloadUrls?.mobile
                         }
-                        label={i18n.t('get.action.label')}
+                        label={i18n.t("get.action.label")}
                         size="small"
                         type="secondary"
                       />
@@ -329,14 +390,14 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
             })}
           </Box>
           {/* spacer */}
-          <Box style={{ marginBottom: '42px' }} />
+          <Box style={{ marginBottom: "42px" }} />
           <Box
             alignItems="center"
             display="flex"
             flexDirection="column"
             gap="36"
             paddingX="36"
-            style={{ textAlign: 'center' }}
+            style={{ textAlign: "center" }}
           >
             <Box
               display="flex"
@@ -345,10 +406,10 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
               textAlign="center"
             >
               <Text color="modalText" size="16" weight="bold">
-                {i18n.t('get.looking_for.title')}
+                {i18n.t("get.looking_for.title")}
               </Text>
               <Text color="modalTextSecondary" size="16">
-                {i18n.t('get.looking_for.mobile.description')}
+                {i18n.t("get.looking_for.mobile.description")}
               </Text>
             </Box>
           </Box>
@@ -363,7 +424,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
       {/* header section */}
       <Box
         background={
-          headerBackgroundContrast ? 'profileForeground' : 'modalBackground'
+          headerBackgroundContrast ? "profileForeground" : "modalBackground"
         }
         display="flex"
         flexDirection="column"
@@ -391,8 +452,8 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
                 alignItems="center"
                 as="button"
                 className={touchableStyles({
-                  active: 'shrinkSm',
-                  hover: 'growLg',
+                  active: "shrinkSm",
+                  hover: "growLg",
                 })}
                 color="accentColor"
                 display="flex"
@@ -400,7 +461,7 @@ export function MobileOptions({ onClose }: { onClose: () => void }) {
                 marginTop="20"
                 onClick={() => setWalletStep(headerBackButtonLink!)}
                 padding="16"
-                style={{ height: 17, willChange: 'transform' }}
+                style={{ height: 17, willChange: "transform" }}
                 transition="default"
                 type="button"
               >