+++ packages/rainbowkit/src/wallets/connectorsForWallets.ts
@@ -1,7 +1,7 @@
-import { CreateConnectorFn } from 'wagmi';
-import { isHexString } from '../utils/colors';
-import { omitUndefinedValues } from '../utils/omitUndefinedValues';
-import { Wallet, WalletList, WalletOptionsParams } from './Wallet';
+import { CreateConnectorFn } from "wagmi";
+import { isHexString } from "../utils/colors";
+import { omitUndefinedValues } from "../utils/omitUndefinedValues";
+import { Wallet, WalletList, WalletOptionsParams } from "./Wallet";
 
 interface WalletListItem extends Wallet {
   index: number;
@@ -9,8 +9,34 @@ interface WalletListItem extends Wallet {
   groupName: string;
 }
 
-export const connectorsForWallets = (
-  walletList: WalletList,
+/*
+  Assemble a list of low-level connectors for use with the
+  `WalletButton` and `RainbowButton` components in custom implementations.
+*/
+export function connectorsForWallets(
+  wallets: Wallet[] | WalletList
+): CreateConnectorFn[];
+
+/*
+  Overload implementation for `connectorsForWallets`.
+  1. Returns a `connectors` function that will then return a Connectors array
+  2. Returns a prepared Connectors array stuffed with dummy WalletList data
+*/
+export function connectorsForWallets(walletList: any) {
+  if ("groupName" in walletList[0]) {
+    return _connectorsForWallets(walletList);
+  } else {
+    return _connectorsForWallets([
+      {
+        groupName: "",
+        wallets: walletList as Wallet[],
+      },
+    ]);
+  }
+}
+
+export const _connectorsForWallets = (
+  walletList: WalletList
 ): CreateConnectorFn[] => {
   let index = -1;
 
@@ -33,7 +59,7 @@ export const connectorsForWallets = (
       // guard against non-hex values for `iconAccent`
       if (wallet?.iconAccent && !isHexString(wallet?.iconAccent)) {
         throw new Error(
-          `Property \`iconAccent\` is not a hex value for wallet: ${wallet.name}`,
+          `Property \`iconAccent\` is not a hex value for wallet: ${wallet.name}`
         );
       }
 
@@ -44,7 +70,7 @@ export const connectorsForWallets = (
         index,
       };
 
-      if (typeof wallet.hidden === 'function') {
+      if (typeof wallet.hidden === "function") {
         potentiallyHiddenWallets.push(walletListItem);
       } else {
         visibleWallets.push(walletListItem);
@@ -61,7 +87,7 @@ export const connectorsForWallets = (
 
   walletListItems.forEach(
     ({ createConnector, groupIndex, groupName, hidden, ...walletMeta }) => {
-      if (typeof hidden === 'function') {
+      if (typeof hidden === "function") {
         // Run the function to check if the wallet needs to be hidden
         const isHidden = hidden();
 
@@ -74,9 +100,9 @@ export const connectorsForWallets = (
       const params = (
         // For now we should only use these as the additional parameters
         additionalParams?: Pick<
-          WalletOptionsParams['rkDetails'],
-          'showQrModal' | 'isWalletConnectModalConnector'
-        >,
+          WalletOptionsParams["rkDetails"],
+          "showQrModal" | "isWalletConnectModalConnector"
+        >
       ) => {
         return {
           rkDetails: omitUndefinedValues({
@@ -89,20 +115,20 @@ export const connectorsForWallets = (
         };
       };
 
-      const isWalletConnectConnector = walletMeta.id === 'walletConnect';
+      const isWalletConnectConnector = walletMeta.id === "walletConnect";
 
       if (isWalletConnectConnector) {
         connectors.push(
           createConnector(
-            params({ showQrModal: true, isWalletConnectModalConnector: true }),
-          ),
+            params({ showQrModal: true, isWalletConnectModalConnector: true })
+          )
         );
       }
 
       const connector = createConnector(params());
 
       connectors.push(connector);
-    },
+    }
   );
 
   return connectors;