+++ packages/rainbowkit/src/wallets/getDefaultWallets.ts
@@ -1,11 +1,12 @@
-import { WalletList } from './Wallet';
-import { connectorsForWallets } from './connectorsForWallets';
-import { braveWallet } from './walletConnectors/braveWallet/braveWallet';
-import { coinbaseWallet } from './walletConnectors/coinbaseWallet/coinbaseWallet';
-import { metaMaskWallet } from './walletConnectors/metaMaskWallet/metaMaskWallet';
-import { rainbowWallet } from './walletConnectors/rainbowWallet/rainbowWallet';
-import { safeWallet } from './walletConnectors/safeWallet/safeWallet';
-import { walletConnectWallet } from './walletConnectors/walletConnectWallet/walletConnectWallet';
+import { CreateConnectorFn } from "wagmi";
+import { WalletList } from "./Wallet";
+import { connectorsForWallets } from "./connectorsForWallets";
+import { braveWallet } from "./walletConnectors/braveWallet/braveWallet";
+import { coinbaseWallet } from "./walletConnectors/coinbaseWallet/coinbaseWallet";
+import { metaMaskWallet } from "./walletConnectors/metaMaskWallet/metaMaskWallet";
+import { rainbowWallet } from "./walletConnectors/rainbowWallet/rainbowWallet";
+import { safeWallet } from "./walletConnectors/safeWallet/safeWallet";
+import { walletConnectWallet } from "./walletConnectors/walletConnectWallet/walletConnectWallet";
 
 export const getDefaultWallets = ({
   appName,
@@ -14,12 +15,12 @@ export const getDefaultWallets = ({
   appName: string;
   projectId: string;
 }): {
-  connectors: ReturnType<typeof connectorsForWallets>;
+  connectors: CreateConnectorFn[];
   wallets: WalletList;
 } => {
   const wallets: WalletList = [
     {
-      groupName: 'Popular',
+      groupName: "Popular",
       wallets: [
         safeWallet(),
         rainbowWallet({ projectId }),