+++ packages/rainbowkit-siwe-next-auth/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rainbow-me/rainbowkit-siwe-next-auth
 
+## 0.3.4
+
+### Patch Changes
+
+- Peer dependency support for RainbowKit `1.3.x`
+
+- Updated dependencies [9ce75a65]
+  - @rainbow-me/rainbowkit@1.3.0
+
+## 0.3.3
+
+### Patch Changes
+
+- 2f56ab23: Adopted `'use client'` directive for safe App Router usage
+- 9dfe0531: Added strict peer incompatibility with `next-auth@5`. RainbowKit will support `next-auth` v5 in a future release with improved App Router support.
+
 ## 0.3.2
 
 ### Patch Changes